Feature - 인벤토리 슬롯 버튼 파라미터 변경 완료

+++ Assets/ScriptableObject/Scripts/AttackStatusData.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class AttackStatusData : DefaultStatusData
 {
     [Header("Attack Info")]
-    public float attack;
-    public float defense;
+    public int attack;
+    public int defense;
     public float critical;
 }
